% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_delete_flow}
\alias{mediaconnect_delete_flow}
\title{Deletes a flow}
\usage{
mediaconnect_delete_flow(FlowArn)
}
\arguments{
\item{FlowArn}{[required] The ARN of the flow that you want to delete.}
}
\description{
Deletes a flow. Before you can delete a flow, you must stop the flow.
}
\section{Request syntax}{
\preformatted{mediaconnect$delete_flow(
  FlowArn = "string"
)
}
}

\keyword{internal}
