% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_start_flow}
\alias{mediaconnect_start_flow}
\title{Starts a flow}
\usage{
mediaconnect_start_flow(FlowArn)
}
\arguments{
\item{FlowArn}{[required] The ARN of the flow that you want to start.}
}
\description{
Starts a flow.
}
\section{Request syntax}{
\preformatted{mediaconnect$start_flow(
  FlowArn = "string"
)
}
}

\keyword{internal}
