% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediaconnect_operations.R
\name{mediaconnect_tag_resource}
\alias{mediaconnect_tag_resource}
\title{Associates the specified tags to a resource}
\usage{
mediaconnect_tag_resource(ResourceArn, Tags)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) of the resource that you want to add tags to.}

\item{Tags}{[required] A map from tag keys to values. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.}
}
\description{
Associates the specified tags to a resource. If the request does not mention an existing tag associated with the resource, that tag is not changed.
}
\section{Request syntax}{
\preformatted{mediaconnect$tag_resource(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

\keyword{internal}
